refactoring day12
