+++ 2019/resources/day12/day12.txt
@@ -1,0 +1,4 @@
+<x=16, y=-8, z=13>
+<x=4, y=10, z=10>
+<x=17, y=-5, z=6>
+<x=13, y=-3, z=0>
