+++ 2019/src/lib.rs
@@ -11,3 +11,4 @@ pub mod day08;
 pub mod day09;
 pub mod day10;
 pub mod day11;
+pub mod day12;
